requirements.txt
PyPDF2
langchain
langchain-google-genai
langchain-community
python-dotenv
google-generativeai
python-multipart
faiss-cpu